fix: removal of .svelte.ts files
