+++ src/routes/api-check/components/alert/+page.svelte
@@ -48,7 +48,7 @@
 <CodeWrapper innerClass="space-y-4">
   <ExampleComponents.Dismissable />
   {#snippet codeblock()}
-    <HighlightCompo codeLang="ts" code={exampleModules["./examples/Dismissable.svelte"] as string} />
+    <HighlightCompo codeLang="ts" code={exampleModules["./examples/dismissable"] as string} />
   {/snippet}
 </CodeWrapper>
