+++ src/routes/api-check/components/alert/examples/index.ts
@@ -3,7 +3,7 @@ export { default as AlertWithList } from "./AlertWithList.svelte";
 export { default as Bordered } from "./Bordered.svelte";
 export { default as BorderAccent } from "./BorderAccent.svelte";
 export { default as Default } from "./Default.svelte";
-export { default as Dismissable } from "./Dismissable.svelte";
+export { default as Dismissable } from "./dismissable";
 export { default as Event } from "./Event.svelte";
 export { default as CustomColor } from "./CustomColor.svelte";
 export { default as Icon } from "./Icon.svelte";
