+++ src/routes/api-check/components/badge/+page.svelte
@@ -90,7 +90,7 @@
 <CodeWrapper innerClass="flex justify-center gap-2">
   <ExampleComponents.Dismissable />
   {#snippet codeblock()}
-    <HighlightCompo codeLang="ts" code={exampleModules["./examples/Dismissable.svelte"] as string} />
+    <HighlightCompo codeLang="ts" code={exampleModules["./examples/dismissable"] as string} />
   {/snippet}
 </CodeWrapper>
