+++ src/routes/api-check/components/badge/examples/index.ts
@@ -1,7 +1,7 @@
 export { default as Bordered } from "./Bordered.svelte";
 export { default as ButtonBadge } from "./ButtonBadge.svelte";
 export { default as Default } from "./Default.svelte";
-export { default as Dismissable } from "./Dismissable.svelte";
+export { default as Dismissable } from "./dismissable";
 export { default as Dismissable2 } from "./Dismissable2.svelte";
 export { default as Dismissable3 } from "./Dismissable3.svelte";
 export { default as Dynamic } from "./Dynamic.svelte";
